更新流程节点查询方式

+++ source/schema/execution.go
@@ -18,7 +18,7 @@ func (Execution) Fields() []ent.Field {
 		field.Int64("proc_def_id").Comment("流程定义id"),
 		field.String("node_infos").MaxLen(5000).Comment("节点审批执行顺序").Optional(),
 		field.Time("start_time").Default(time.Now()).Comment("开始时间").Optional(),
-		field.Int8("is_del").Default(0).Comment("是否删除,0:未删除,1:已删除").Optional(),
+		field.Int32("is_del").Default(0).Comment("是否删除,0:未删除,1:已删除").Optional(),
 		field.Time("create_time").Default(time.Now()).Comment("创建时间").Optional(),
 		field.Time("update_time").Default(time.Now()).Comment("流程修改时间").Optional(),
 	}
